Revert README helper section
